backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
pytz==2024.2
redis==5.2.1
sqlparse==0.5.2
typing_extensions==4.12.2
